Move button to components package (#6173)
* Move button to components package

**Why**: Preparatory work ahead of upcoming FlowStateMachine v2 work, where it's expected we will want to use this component in packages outside document-capture.

[skip changelog]

* Convert Button to TypeScript
